requirements.txt
python-dotenv==1.2.1
fastapi==0.128.8
uvicorn==0.39.0
sqlalchemy==2.0.49
greenlet==2.0.2
psycopg2-binary==2.9.11
openai==2.32.0
pandas==2.3.3
gspread==6.2.1
pytest==8.4.2
